[NF] fetching and mapping for fetch options

+++ common/model/db/RawOptionsDataDto.py
@@ -1,4 +1,4 @@
-from sqlalchemy import DateTime, Column, Integer, String
+from sqlalchemy import Column, String, BigInteger, Boolean
 from sqlalchemy.dialects.postgresql import TIMESTAMP, JSON
 
 from common.model.db.BaseDto import BaseDto
@@ -7,7 +7,7 @@ from common.model.db.BaseDto import BaseDto
 class RawOptionsDataDto(BaseDto):
     __tablename__ = "rawdata"
 
-    id: int = Column(Integer, primary_key=True)
+    id: int = Column(BigInteger, primary_key=True)
     raw_data: str = Column(JSON)
     writer: str = Column(String)
     datetime: str = Column(TIMESTAMP(timezone=True))
